+++ packages/nimble-blazor/Tests/NimbleBlazor.Tests/Unit/Components/NimbleTabsTests.cs
@@ -1,4 +1,5 @@
 ﻿using System.Linq;
+using System.Threading.Tasks;
 using AngleSharp.Dom;
 using Bunit;
 using Xunit;
@@ -31,6 +32,29 @@ public class NimbleTabsTests
         Assert.Equal(expectedChildrenNames, actualChildNodeNames);
     }
 
+    [Fact]
+    public async void NimbleTabs_ChangeActiveIdToNull_ActiveTabIdDoesNotSwitch()
+    {
+        var expectedActiveTabId = "tab1";
+        var tabs = CreateTwoTabsWithActiveTabIdSet(expectedActiveTabId);
+
+        await TriggerNimbleTabsActiveIdChangeEventAsync(tabs, eventArgs: null);
+
+        Assert.Equal(expectedActiveTabId, tabs.Instance.ActiveId);
+    }
+
+    [Fact]
+    public async void NimbleTabs_ChangeActiveId_ActiveTabUpdates()
+    {
+        var expectedActiveTabId = "tab2";
+        var eventArgs = new TabsChangeEventArgs { ActiveId = expectedActiveTabId };
+        var tabs = CreateTwoTabsWithActiveTabIdSet(activeTabId: "tab1");
+
+        await TriggerNimbleTabsActiveIdChangeEventAsync(tabs, eventArgs);
+
+        Assert.Equal(expectedActiveTabId, tabs.Instance.ActiveId);
+    }
+
     private IRenderedComponent<NimbleTabs> RenderTabsWithContent()
     {
         var context = new TestContext();
@@ -43,5 +67,32 @@ public class NimbleTabsTests
         parameters.AddChildContent<NimbleTab>();
         parameters.AddChildContent<NimbleTabPanel>();
         parameters.AddChildContent<NimbleTabsToolbar>();
+    }
+
+    private IRenderedComponent<NimbleTabs> CreateTwoTabsWithActiveTabIdSet(string activeTabId)
+    {
+        var context = new TestContext();
+        context.JSInterop.Mode = JSRuntimeMode.Loose;
+        var tabComponent = context.RenderComponent<NimbleTabs>(parameters =>
+        {
+            parameters.Add(x => x.ActiveId, activeTabId);
+            parameters.AddChildContent<NimbleTab>();
+            parameters.AddChildContent<NimbleTab>();
+            parameters.AddChildContent<NimbleTabPanel>();
+            parameters.AddChildContent<NimbleTabPanel>();
+        });
+
+        Assert.Equal(activeTabId, tabComponent.Instance.ActiveId);
+
+        return tabComponent;
+    }
+
+    private async Task TriggerNimbleTabsActiveIdChangeEventAsync(
+        IRenderedComponent<NimbleTabs> tabs,
+        TabsChangeEventArgs eventArgs)
+    {
+        var tabsElement = tabs.Find("nimble-tabs");
+
+        await tabsElement.TriggerEventAsync("onnimbletabsactiveidchange", eventArgs);
     }
 }