+++ packages/nimble-components/src/menu/index.ts
@@ -5,13 +5,9 @@ import {
 } from '@microsoft/fast-foundation';
 import { styles } from './styles';
 
-// FAST menu item uses DesignSystem.tagFor(anchoredRegion) to position nested menus
-// https://github.com/microsoft/fast/blob/3ad50c8eec98b3156a4559586ca76e4abebefad8/packages/web-components/fast-foundation/src/menu-item/menu-item.template.ts#L83
-// This needs to be able to find the Nimble anchored region even if the application doesn't import it explicitly or via all-components
-// This may become unnecessary with fast-foundation vNext which no longer uses anchored region
-// https://github.com/microsoft/fast/pull/6457
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { AnchoredRegion } from '../anchored-region';
+// FAST menu template requires an anchored region is available using tagFor DI
+// Register anchored region explicitly to make sure it is defined for the template
+import '../anchored-region';
 
 declare global {
     interface HTMLElementTagNameMap {